Split rpc functions into new file.
