+++ tests/basicswap/test_run.py
@@ -37,8 +37,11 @@ from basicswap.basicswap import (
 from basicswap.util import (
     COIN,
     toWIF,
-    callrpc_cli,
     dumpje,
+)
+from basicswap.rpc import (
+    callrpc_cli,
+    waitForRPC,
 )
 from basicswap.key import (
     ECKey,
@@ -208,17 +211,6 @@ def run_loop(self):
             c.update()
         ltcRpc('generatetoaddress 1 {}'.format(self.ltc_addr))
         btcRpc('generatetoaddress 1 {}'.format(self.btc_addr))
-
-
-def waitForRPC(rpc_func, wallet=None):
-    for i in range(5):
-        try:
-            rpc_func('getwalletinfo')
-            return
-        except Exception as ex:
-            logging.warning('Can\'t connect to daemon RPC: %s.  Trying again in %d second/s.', str(ex), (1 + i))
-            time.sleep(1 + i)
-    raise ValueError('waitForRPC failed')
 
 
 def checkForks(ro):
